Now use a hashTable
Using a hash table means we can only upload files that are new or
modified. The hashing relies a lot on the XML always being outputted in
the same way, the conversion to JSON happening the same way each time
and the phase of the moon (probably).

Which is to say, it’s good enough at the moment, we should figure out a
better way.